WIP: Convert Account Counters to typescript
